location/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.2.2
apispec==6.8.1
apispec-webframeworks==1.2.0 
marshmallow==3.26.1 
flask_swagger_ui==4.11.1
